+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed
+### Fixed
 
-- ???
+- Fixed the issue that `RexsVersion.findByModelVersion(...)` did not searches for additional model versions.
 
 
 ## [0.11.0] - 2024-12-05